refactor: 렌더링 함수 별도 폴더로 분리
